requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
decorator==5.2.1
exceptiongroup==1.3.0
fastapi==0.115.12
ffmpeg-python==0.2.0
future==1.0.0
h11==0.16.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
numpy==2.0.2
pillow==10.4.0
proglog==0.1.12
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.1.0
python-multipart==0.0.20
requests==2.32.3
sniffio==1.3.1
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
